requirements.txt
fastapi
inngest
llama-index-core
llama-index-readers-file
python-dotenv
qdrant-client
uvicorn
streamlit
openai
pydantic>=2.0
qdrant_client
llama_index